Refactor token verification as part of access-account domain
