+++ api/src/domain/access/index.ts
@@ -40,3 +40,13 @@ export async function registerAccountWithPassword(
   const account = await accountRepository.create(email, passwordHash);
   return accountRepository.saveAccount(account);
 }
+
+export type VerifyTokenDI = {
+  tokenizer: Tokenizer;
+};
+
+export async function verifyToken(di: VerifyTokenDI, token: string) {
+  const { tokenizer } = di;
+  const tokenPayload = await tokenizer.readToken(token);
+  return tokenPayload;
+}
