--- api/src/domain/services/verify-token-payload.logic.ts
@@ -1,11 +1,0 @@
-import { Tokenizer } from '../access/tools/tokenizer.tool';
-
-export type VerifyTokenDI = {
-  tokenizer: Tokenizer;
-};
-
-export async function verifyToken(di: VerifyTokenDI, token: string) {
-  const { tokenizer } = di;
-  const tokenPayload = await tokenizer.readToken(token);
-  return tokenPayload;
-}
